+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -13,9 +13,9 @@ import java.util.stream.Collectors;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.StringUtil;
-import seedu.address.model.person.Name;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
+import seedu.address.model.task.Name;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes
