+++ src/test/java/guitests/EditCommandTest.java
@@ -8,8 +8,8 @@ import org.junit.Test;
 import guitests.guihandles.TaskCardHandle;
 import seedu.address.commons.core.Messages;
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.model.person.Name;
 import seedu.address.model.tag.Tag;
+import seedu.address.model.task.Name;
 import seedu.address.testutil.TaskBuilder;
 import seedu.address.testutil.TestTask;
