+++ src/test/java/seedu/address/storage/XmlAddressBookStorageTest.java
@@ -15,7 +15,7 @@ import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.model.ReadOnlyToDoList;
 import seedu.address.model.ToDoList;
-import seedu.address.model.person.Task;
+import seedu.address.model.task.Task;
 import seedu.address.testutil.TypicalTestTasks;
 
 public class XmlAddressBookStorageTest {
