+++ src/test/java/seedu/address/testutil/AddressBookBuilder.java
@@ -2,9 +2,9 @@ package seedu.address.testutil;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.ToDoList;
-import seedu.address.model.person.Task;
-import seedu.address.model.person.UniqueTaskList;
 import seedu.address.model.tag.Tag;
+import seedu.address.model.task.Task;
+import seedu.address.model.task.UniqueTaskList;
 
 /**
  * A utility class to help with building Addressbook objects.
